# Node16 shrinking constructor: remove the child-to-be-removed before t…